基类static成员变量不能在基类构造函数中new生成，否则子类都会new生成这个static成员。win10没有db访问权限，无法写入数据库

+++ ItshfSim/sql/simsql.cpp
@@ -27,6 +27,11 @@ SimSql::~SimSql(void)
 // 插入数据
 int SimSql::insert(int tab, const Time* ts, int valid, int glbChId, int snr, int noise)
 {
+    if (m_handle == nullptr) {
+        printf("err: null DB handle\n");
+        return SQLITE_ERROR;
+    }
+
     char* errMsg = nullptr;
     const char* tlist[] = {"SCAN", "LINK"};
     char* sql = sqlite3_mprintf("insert into %s values(%d,%d,%d,%d,%d, %d,%d,%d,%d,%d)", tlist[tab],
@@ -45,6 +50,11 @@ int SimSql::insert(int tab, const Time* ts, int valid, int glbChId, int snr, int
 // 筛选数据
 int SimSql::select(int tab, const Time* ts, int min, QList<FreqInfo>& list)
 {
+    if (m_handle == nullptr) {
+        printf("err: null DB handle\n");
+        return SQLITE_ERROR;
+    }
+
     int minDay, minHr, minMin;
     int maxDay, maxHr, maxMin;
 
@@ -97,6 +107,11 @@ int SimSql::select(int tab, const Time* ts, int min, QList<FreqInfo>& list)
 // 丢弃所有数据
 void SimSql::drop(int tab)
 {
+    if (m_handle == nullptr) {
+        printf("err: null DB handle\n");
+        return;
+    }
+
     char* errMsg = nullptr;
     const char* tlist[] = {"SCAN", "LINK"};
     char* sql = sqlite3_mprintf("drop table if exists %s", tlist[tab]);
